service: logging to syslog
