+++ pkg/util/util.go
@@ -4,9 +4,13 @@ import (
 	"os"
 
 	jsoniter "github.com/json-iterator/go"
+	"github.com/nezhahq/service"
 )
 
-var Json = jsoniter.ConfigCompatibleWithStandardLibrary
+var (
+	Json   = jsoniter.ConfigCompatibleWithStandardLibrary
+	Logger service.Logger
+)
 
 func IsWindows() bool {
 	return os.PathSeparator == '\\' && os.PathListSeparator == ';'
